pyproject.toml
[project]
name = "socketry"
version = "0.1.0"
description = "Python API and CLI for controlling Jackery portable power stations"
readme = "README.md"
authors = [
    { name = "Jesus Lopez", email = "jesus@jesusla.com" }
]
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "typer>=0.9",
    "paho-mqtt>=2.0",
    "requests>=2.31",
    "pycryptodome>=3.19",
]

[project.scripts]
socketry = "socketry.cli:app"

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.14",
    "types-requests>=2.31.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
ignore_missing_imports = false
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--cov=socketry",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=json",
]

[tool.coverage.run]
source = ["src/socketry"]
omit = ["tests/*", "**/__pycache__/*"]

[tool.coverage.report]
fail_under = 20
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
